+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/room/failure/CreateRoomFailure.kt
@@ -22,4 +22,9 @@ import org.matrix.android.sdk.api.failure.MatrixError
 sealed class CreateRoomFailure : Failure.FeatureFailure() {
     object CreatedWithTimeout : CreateRoomFailure()
     data class CreatedWithFederationFailure(val matrixError: MatrixError) : CreateRoomFailure()
+    sealed class RoomAliasError : CreateRoomFailure() {
+        object AliasEmpty : RoomAliasError()
+        object AliasNotAvailable : RoomAliasError()
+        object AliasInvalid : RoomAliasError()
+    }
 }