+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/create/CreateRoomTask.kt
@@ -31,8 +31,10 @@ import org.matrix.android.sdk.internal.database.model.RoomEntityFields
 import org.matrix.android.sdk.internal.database.model.RoomSummaryEntity
 import org.matrix.android.sdk.internal.database.query.where
 import org.matrix.android.sdk.internal.di.SessionDatabase
+import org.matrix.android.sdk.internal.di.UserId
 import org.matrix.android.sdk.internal.network.executeRequest
 import org.matrix.android.sdk.internal.session.room.RoomAPI
+import org.matrix.android.sdk.internal.session.room.alias.RoomAliasDescription
 import org.matrix.android.sdk.internal.session.room.read.SetReadMarkersTask
 import org.matrix.android.sdk.internal.session.user.accountdata.DirectChatsHelper
 import org.matrix.android.sdk.internal.session.user.accountdata.UpdateUserAccountDataTask
@@ -45,6 +47,7 @@ internal interface CreateRoomTask : Task<CreateRoomParams, String>
 
 internal class DefaultCreateRoomTask @Inject constructor(
         private val roomAPI: RoomAPI,
+        @UserId private val userId: String,
         @SessionDatabase private val monarchy: Monarchy,
         private val directChatsHelper: DirectChatsHelper,
         private val updateUserAccountDataTask: UpdateUserAccountDataTask,
@@ -61,6 +64,31 @@ internal class DefaultCreateRoomTask @Inject constructor(
                     ?: throw IllegalStateException("You can't create a direct room without an invitedUser")
         } else null
 
+        if (params.preset == CreateRoomPreset.PRESET_PUBLIC_CHAT) {
+            if (params.roomAliasName.isNullOrEmpty()) {
+                throw CreateRoomFailure.RoomAliasError.AliasEmpty
+            }
+            // Check alias availability
+            val fullAlias = "#" + params.roomAliasName + ":" + userId.substringAfter(":")
+            try {
+                executeRequest<RoomAliasDescription>(eventBus) {
+                    apiCall = roomAPI.getRoomIdByAlias(fullAlias)
+                }
+            } catch (throwable: Throwable) {
+                if (throwable is Failure.ServerError && throwable.httpCode == 404) {
+                    // This is a 404, so the alias is available: nominal case
+                    null
+                } else {
+                    // Other error, propagate it
+                    throw throwable
+                }
+            }
+                    ?.let {
+                        // Alias already exists: error case
+                        throw CreateRoomFailure.RoomAliasError.AliasNotAvailable
+                    }
+        }
+
         val createRoomBody = createRoomBodyBuilder.build(params)
 
         val createRoomResponse = try {
@@ -68,14 +96,18 @@ internal class DefaultCreateRoomTask @Inject constructor(
                 apiCall = roomAPI.createRoom(createRoomBody)
             }
         } catch (throwable: Throwable) {
-            if (throwable is Failure.ServerError
-                    && throwable.httpCode == 403
-                    && throwable.error.code == MatrixError.M_FORBIDDEN
-                    && throwable.error.message.startsWith("Federation denied with")) {
-                throw CreateRoomFailure.CreatedWithFederationFailure(throwable.error)
-            } else {
-                throw throwable
+            if (throwable is Failure.ServerError) {
+                if (throwable.httpCode == 403
+                        && throwable.error.code == MatrixError.M_FORBIDDEN
+                        && throwable.error.message.startsWith("Federation denied with")) {
+                    throw CreateRoomFailure.CreatedWithFederationFailure(throwable.error)
+                } else if (throwable.httpCode == 400
+                        && throwable.error.code == MatrixError.M_UNKNOWN
+                        && throwable.error.message == "Invalid characters in room alias") {
+                    throw CreateRoomFailure.RoomAliasError.AliasInvalid
+                }
             }
+            throw throwable
         }
         val roomId = createRoomResponse.roomId
         // Wait for room to come back from the sync (but it can maybe be in the DB if the sync response is received before)