+++ CHANGES.md
@@ -10,6 +10,7 @@ Improvements 🙌:
  - Use RoomMember instead of User in the context of a Room.
  - Ask for explicit user consent to send their contact details to the identity server (#2375)
  - Handle events of type "m.room.server_acl" (#890)
+ - Room creation form: add advanced section to disable federation (#1314)
  - Move "Enable Encryption" from room setting screen to room profile screen (#2394)
 
 Bugfix 🐛: